requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
Flask-Cors>=4.0
python-dotenv>=1.0
PyMySQL>=1.1      # 仅当你要连 MySQL 时需要
EOF